+++ bicleaner/bicleaner_classifier_full.py
@@ -21,7 +21,7 @@ from multiprocessing import Queue, Process, Value, cpu_count
 from tempfile import NamedTemporaryFile, gettempdir
 from timeit import default_timer
 from toolwrapper import ToolWrapper
-from mosestokenizer import MosesTokenizer
+from sacremoses import MosesTokenizer
 
 #Allows to load modules while inside or outside the package
 try:
@@ -253,11 +253,11 @@ def classifier_process(i, jobs_queue, output_queue, args):
     if args.source_tokeniser_path:    
         source_tokeniser = ToolWrapper(args.source_tokeniser_path.split(' '))
     else:
-        source_tokeniser = MosesTokenizer(args.source_lang)
+        source_tokeniser = MosesTokenizer(lang=args.source_lang)
     if args.target_tokeniser_path:
         target_tokeniser = ToolWrapper(args.target_tokeniser_path.split(' '))
     else:
-        target_tokeniser = MosesTokenizer(args.target_lang)
+        target_tokeniser = MosesTokenizer(lang=args.target_lang)
         
     '''
     #Load LM for fluency scoring
@@ -314,8 +314,7 @@ def classifier_process(i, jobs_queue, output_queue, args):
                     if sl_sentence and tl_sentence and len(sl_sentence.strip()) != 0 and len(tl_sentence.strip()) != 0 and (args.disable_hardrules or  wrong_tu(sl_sentence.strip(),tl_sentence.strip(), args, lm_filter, porn_removal, tokenizer)== False):
                         #if disable_hardrules == 1 --> the second part (and) is always true
                         features = feature_extract(sl_sentence, tl_sentence, source_tokeniser, target_tokeniser, args)
-
-
+                        
                         feats.append([float(v) for v in features])
                         valid_sentences.append(True)
                     else: